requirements-common.txt
########################################
# Genesis Common Requirements
########################################

# API service clients

## TiDB Vector Search and Embedding
pytidb[models]==0.0.13

## weaviate
weaviate-client==4.16.10

## PostgreSQL driver
psycopg==3.2.10

## OpenAI Platform
openai==1.103.0

## Ollama Python client (connects to host or service via OLLAMA_BASE_URL)
ollama==0.5.3

## LangChain/Graph ecosystem and connectors
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
langchain-openai==0.3.32
langchain-ollama==0.3.7
langchain-google-genai==2.1.10
langchain-text-splitters==0.3.10
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.20

## Hugging Face / tokenization
huggingface-hub==0.34.4
datasets==3.6.0
tokenizers==0.22.0
tiktoken==0.11.0


# General libs (local execution)

## Embeddings / ML utils
sentence-transformers==5.1.0
numpy==2.2.6
pandas==2.3.2
scipy==1.16.1
scikit-learn==1.7.1

## Computer vision (base, CPU/GPU-specific extras handled separately)
opencv-python==4.10.0.84

## Paddle ecosystem clients that rely on PaddlePaddle
paddleocr==3.1.0
paddlex==3.1.3

## Core runtime and assorted utilities
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.6
annotated-types==0.7.0
appdirs==1.4.4
attrs==25.3.0
audioread==3.0.1
bce-python-sdk==0.9.45
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==5.2.1
dill==0.3.8
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.13.1
filetype==1.2.0
flet==0.28.3
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.1
marisa-trie==1.3.1
MarkupSafe==2.1.5
marshmallow==3.26.1
modelscope==1.29.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
mysqlclient==2.2.7
networkx==3.3
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.0.0
platformdirs==4.4.0
pooch==1.8.2
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyHyphen==4.0.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
SQLAlchemy==2.0.43
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0

# Backend API (FastAPI)
fastapi==0.116.1
uvicorn[standard]==0.35.0
alembic==1.12.1
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-magic==0.4.27